Features for tracking elements, document code.

+++ nCubed.EFCore.Test/Fakes/Entities/ProjectDetail.cs
@@ -5,7 +5,7 @@ namespace nCubed.EFCore.Test.Entities
 {
     public class ProjectDetail
     {
-        public int ProjectId { get; set; }
+        public long ProjectId { get; set; }
         public Project Project { get; set; }
         public decimal Budget { get; set; }
         public bool Critical { get; set; }
