+++ nCubed.EFCore.Test/Fakes/Entities/Resource.cs
@@ -6,7 +6,7 @@ namespace nCubed.EFCore.Test.Entities
 {
     public class Resource
     {
-        public int ResourceId { get; set; }
+        public long ResourceId { get; set; }
         public string Name { get; set; }
         public ContactInformation ContactInformation { get; set; }
         public ICollection<ProjectResource> ProjectResources { get; private set; } = new HashSet<ProjectResource>();
